Refactor endpoint catch runtime helpers
